- circular reference fix
